backend/requirements.txt
annotated-doc
annotated-types
anyio
aiohttp
certifi
chardet
charset-normalizer
click
coverage
dulwich
fastapi
filelock
fsspec
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
idna
iniconfig
Jinja2
joblib
MarkupSafe
mpmath
networkx
numpy
ollama
packaging
pillow
pluggy
pydantic
pydantic_core
pdfplumber
PyPDF2
python-docx
openpyxl
python-pptx
python-multipart
Pygments
pytest
pytest-cov
python-dotenv
PyYAML
qdrant-client
regex
requests
SQLAlchemy
psycopg2-binary
pymysql
psutil
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sniffio
starlette
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
trafilatura
typing-inspection
typing_extensions
urllib3
uvicorn[standard]
watchfiles
watchdog
websockets
SpeechRecognition
pydub
moviepy
ffmpeg-python
schedule
filelock==3.13.1
mcp>=1.2.0
fastmcp>=0.1.0
pytest-asyncio>=0.23.0
